requirements.txt
python-stdnum >=1.17,<2.0
regex >= 2021.11.10

pii-data >= 0.2.0, <1.0.0
pii-extract-base >= 0.0.1, <1.0.0
